Moved code for additional RI tests from separate GtDefaultPersoTestRi to parent test GtDefaultPersoTest, deleted the now obsolete test and the respective launch config.
