--- de.persosim.simulator.test/launch/GtDefaultPersoTestRi.launch
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<launchConfiguration type="org.eclipse.jdt.junit.launchconfig">
-<listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
-<listEntry value="/de.persosim.simulator.test/gtTests/de/persosim/simulator/test/globaltester/perso/GtDefaultPersoTestRi.java"/>
-</listAttribute>
-<listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
-<listEntry value="1"/>
-</listAttribute>
-<mapAttribute key="org.eclipse.debug.core.preferred_launchers">
-<mapEntry key="[debug]" value="org.eclipse.jdt.junit.launchconfig"/>
-<mapEntry key="[run]" value="org.eclipse.jdt.junit.launchconfig"/>
-</mapAttribute>
-<stringAttribute key="org.eclipse.jdt.junit.CONTAINER" value=""/>
-<booleanAttribute key="org.eclipse.jdt.junit.KEEPRUNNING_ATTR" value="false"/>
-<stringAttribute key="org.eclipse.jdt.junit.TESTNAME" value=""/>
-<stringAttribute key="org.eclipse.jdt.junit.TEST_KIND" value="org.eclipse.jdt.junit.loader.junit4"/>
-<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="de.persosim.simulator.test.globaltester.perso.GtDefaultPersoTestRi"/>
-<stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="de.persosim.simulator.test"/>
-<stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value="-javaagent:${project_loc:/de.persosim.simulator.test}/lib/jmockit.jar&#13;&#10;-Dcoverage-srcDirs=${project_loc:/de.persosim.simulator}/src,${project_loc:/de.persosim.simulator.test}/src,${project_loc:/de.persosim.simulator.test}/gtTests"/>
-</launchConfiguration>
